+++ .php-cs-fixer.dist.php
@@ -1,0 +1,242 @@
+<?php
+
+// Last reviewed: v2.19 (Testament)
+// Based on https://gist.github.com/laravel-shift/cab527923ed2a109dda047b97d53c200
+
+use PhpCsFixer\Config;
+use PhpCsFixer\Finder;
+
+$rules = [
+    '@PSR1' => true,
+    '@PSR2' => true,
+    '@PhpCsFixer' => true,
+    '@Symfony' => true,
+    '@PHP70Migration' => true,
+    '@PHP70Migration:risky' => true,
+    '@PHP71Migration' => true,
+    '@PHP71Migration:risky' => true,
+    '@PHP73Migration' => true,
+    '@PHPUnit75Migration:risky' => true,
+    'final_class' => false,
+    'new_with_braces' => false,
+    'strict_comparison' => true,
+    'list_syntax' => ['syntax' => 'short'],
+    'mb_str_functions' => true,
+    'class_attributes_separation' => [
+        'elements' => [
+            'method' => 'one',
+        ],
+    ],
+    'no_extra_blank_lines' => [
+        'tokens' => [
+            'break',
+            'continue',
+            'curly_brace_block',
+            'extra',
+            'parenthesis_brace_block',
+            'return',
+            'square_brace_block',
+            'throw',
+            'use',
+            'use_trait',
+            'switch',
+
+            'case',
+            'default',
+        ],
+    ],
+    'no_blank_lines_before_namespace' => false,
+    'nullable_type_declaration_for_default_null_value' => true,
+    'increment_style' => ['style' => 'pre'],
+    'self_static_accessor' => true,
+    'static_lambda' => false,
+    'no_empty_phpdoc' => true,
+    'no_superfluous_phpdoc_tags' => [
+        'remove_inheritdoc' => true,
+    ],
+    'phpdoc_line_span' => [
+        'const' => 'multi',
+        'method' => 'multi',
+        'property' => 'multi',
+    ],
+    'general_phpdoc_tag_rename' => true,
+    'phpdoc_add_missing_param_annotation' => ['only_untyped' => true],
+    'phpdoc_align' => ['align' => 'left'],
+    'phpdoc_indent' => true,
+    'phpdoc_inline_tag_normalizer' => true,
+    'phpdoc_no_access' => true,
+    'phpdoc_no_empty_return' => false,
+    'phpdoc_no_package' => true,
+    'phpdoc_no_useless_inheritdoc' => true,
+    'phpdoc_order' => true,
+    'phpdoc_order_by_value' => false,
+    'phpdoc_scalar' => true,
+    'phpdoc_single_line_var_spacing' => true,
+    'phpdoc_summary' => true,
+    'phpdoc_tag_casing' => ['tags' => ['inheritDoc']],
+    'phpdoc_tag_type' => true,
+    'phpdoc_to_comment' => true,
+    'phpdoc_trim' => true,
+    'phpdoc_types' => ['groups' => ['simple', 'alias']],
+    'phpdoc_types_order' => ['null_adjustment' => 'always_last'],
+    'phpdoc_var_annotation_correct_order' => true,
+    'phpdoc_var_without_name' => true,
+    'align_multiline_comment' => ['comment_type' => 'phpdocs_like'],
+    'php_unit_test_class_requires_covers' => false,
+    'php_unit_internal_class' => false,
+    'yoda_style' => false,
+    'ordered_class_elements' => [
+        'order' => [
+            'use_trait',
+            'constant_public',
+            'constant_protected',
+            'constant_private',
+
+            'property_static',
+            'property_public_static',
+            'property_protected_static',
+            'property_private_static',
+
+            'property',
+            'property_public',
+            'property_protected',
+            'property_private',
+
+            'construct',
+            'destruct',
+            'magic',
+
+            'method_static',
+            'method_public_abstract_static',
+            'method_public_static',
+            'method_protected_abstract_static',
+            'method_protected_static',
+            'method_private_static',
+            'method_public_abstract',
+            'method_public',
+            'method_protected_abstract',
+            'method_protected',
+            'method_private',
+        ],
+    ],
+    'array_syntax' => ['syntax' => 'short'],
+    'blank_line_after_namespace' => true,
+    'blank_line_after_opening_tag' => true,
+    'braces' => true,
+    'cast_spaces' => true,
+    'concat_space' => [
+        'spacing' => 'none',
+    ],
+    'declare_equal_normalize' => true,
+    'elseif' => true,
+    'encoding' => true,
+    'full_opening_tag' => true,
+    'fully_qualified_strict_types' => true, // added by Shift
+    'function_declaration' => true,
+    'function_typehint_space' => true,
+    'heredoc_to_nowdoc' => true,
+    'include' => true,
+    'indentation_type' => true,
+    'linebreak_after_opening_tag' => true,
+    'line_ending' => true,
+    'lowercase_cast' => true,
+    'lowercase_keywords' => true,
+    'lowercase_static_reference' => true, // added from Symfony
+    'magic_method_casing' => true, // added from Symfony
+    'magic_constant_casing' => true,
+    'method_argument_space' => true,
+    'native_function_casing' => true,
+    'no_alias_functions' => false,
+    'no_blank_lines_after_class_opening' => true,
+    'no_blank_lines_after_phpdoc' => true,
+    'no_closing_tag' => true,
+    'no_empty_statement' => true,
+    'no_leading_import_slash' => true,
+    'no_leading_namespace_whitespace' => true,
+    'no_mixed_echo_print' => [
+        'use' => 'echo',
+    ],
+    'no_multiline_whitespace_around_double_arrow' => true,
+    'multiline_whitespace_before_semicolons' => [
+        'strategy' => 'no_multi_line',
+    ],
+    'no_short_bool_cast' => true,
+    'no_singleline_whitespace_before_semicolons' => true,
+    'no_spaces_after_function_name' => true,
+    'no_spaces_inside_parenthesis' => true,
+    'no_trailing_comma_in_list_call' => false,
+    'no_trailing_comma_in_singleline_array' => true,
+    'no_trailing_whitespace' => true,
+    'no_trailing_whitespace_in_comment' => true,
+    'no_unreachable_default_argument_value' => true,
+    'no_useless_return' => true,
+    'no_whitespace_before_comma_in_array' => true,
+    'no_whitespace_in_blank_line' => true,
+    'normalize_index_brace' => true,
+    'not_operator_with_successor_space' => true,
+    'object_operator_without_whitespace' => true,
+    'self_accessor' => false,
+    'short_scalar_cast' => true,
+    'simplified_null_return' => false, // disabled by Shift
+    'single_blank_line_at_eof' => true,
+    'single_blank_line_before_namespace' => true,
+    'single_import_per_statement' => true,
+    'single_line_after_imports' => true,
+    'single_line_comment_style' => [
+        'comment_types' => ['hash'],
+    ],
+    'single_quote' => true,
+    'space_after_semicolon' => true,
+    'standardize_not_equals' => true,
+    'switch_case_semicolon_to_colon' => true,
+    'switch_case_space' => true,
+    'ternary_operator_spaces' => true,
+    'trim_array_spaces' => true,
+    'unary_operator_spaces' => true,
+    'whitespace_after_comma_in_array' => true,
+    'constant_case' => ['case' => 'lower'],
+    'psr_autoloading' => true,
+    'trailing_comma_in_multiline' => ['elements' => ['arrays']],
+    'binary_operator_spaces' => [
+        'default' => 'single_space',
+    ],
+    'types_spaces' => [
+        'space' => 'none',
+    ],
+    'blank_line_before_statement' => [
+        'statements' => ['return'],
+    ],
+    'class_definition' => [
+        'multi_line_extends_each_single_line' => true,
+        'single_item_single_line' => true,
+        'single_line' => true,
+    ],
+    'ordered_imports' => [
+        'sort_algorithm' => 'alpha',
+    ],
+    'no_unneeded_control_parentheses' => [
+        'statements' => ['break', 'clone', 'continue', 'echo_print', 'return', 'switch_case', 'yield'],
+    ],
+    'no_spaces_around_offset' => [
+        'positions' => ['inside', 'outside'],
+    ],
+    'visibility_required' => [
+        'elements' => ['property', 'method', 'const'],
+    ],
+];
+
+$finder = Finder::create()
+    ->in([
+        __DIR__.'/src',
+        __DIR__.'/tests',
+    ])
+    ->name('*.php')
+    ->notName('*.blade.php')
+    ->ignoreDotFiles(true)
+    ->ignoreVCS(true);
+
+return (new Config)
+    ->setRules($rules)
+    ->setFinder($finder)
+    ->setRiskyAllowed(true)
+    ->setUsingCache(true);